requirements.txt
# Python dependencies for OpenClaw Agent

# Core
python-dotenv>=1.0.0
aiohttp>=3.9.0

# Ollama client
ollama>=0.1.0

# Google Gemini
google-generativeai>=0.3.0

# Pydantic for settings
pydantic>=2.5.0
pydantic-settings>=2.1.0

# WhatsApp integration (Python side for message handling)
flask>=3.0.0
flask-cors>=4.0.0

# Utilities
loguru>=0.7.0
tenacity>=8.2.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
